chore: extract custom hub data creator into helper
