+++ utils/config/types.go
@@ -1,0 +1,6 @@
+package config
+
+type PathValue struct {
+	Path  string
+	Value interface{}
+}
